+++ manul/src/protocol.rs
@@ -4,8 +4,8 @@ API for protocol implementors.
 A protocol is a directed acyclic graph with the nodes being objects of types implementing [`Round`]
 (to be specific, "acyclic" means that the values returned by [`Round::id`]
 should not repeat during the protocol execution; the types might).
-The starting point is a type also implementing [`FirstRound`].
-All the rounds should have their associated type [`Round::Protocol`] set to the same [`Protocol`] instance
+The starting point is a type that implements [`FirstRound`].
+All the rounds must have their associated type [`Round::Protocol`] set to the same [`Protocol`] instance
 to be executed by a [`Session`](`crate::session::Session`).
 
 For more details, see the documentation of the mentioned traits.
@@ -15,7 +15,6 @@ mod errors;
 mod object_safe;
 mod round;
 
-pub use crate::session::SessionId;
 pub use errors::{
     DeserializationError, DirectMessageError, EchoBroadcastError, FinalizeError, LocalError, MessageValidationError,
     ProtocolValidationError, ReceiveError, RemoteError,