+++ manul/src/session.rs
@@ -1,7 +1,10 @@
 /*!
 API for protocol users.
 
-This is some description.
+The round-based protocols `manul` is designed to build use a [`Session`] object to drive the protocol forward.
+Each participant constructs a [`Session`], defining the actions needed for each round (who to send messages
+to, what kind of message and what to do next etc). The rest of the API from this module provide auxilliary
+types: setup and parametrization, errors and outcomes.
 */
 
 mod echo;